+++ packages/article-list/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.19.31](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.19.30...@times-components/article-list@9.19.31) (2024-12-16)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
+## [9.19.30](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.19.29...@times-components/article-list@9.19.30) (2024-12-16)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
+## [9.19.29](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.19.28...@times-components/article-list@9.19.29) (2024-12-13)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
 ## [9.19.28](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.19.27...@times-components/article-list@9.19.28) (2024-12-05)
 
 **Note:** Version bump only for package @times-components/article-list