+++ packages/article-main-comment/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.93.7](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.93.6...@times-components/article-main-comment@2.93.7) (2024-12-17)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.93.6](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.93.5...@times-components/article-main-comment@2.93.6) (2024-12-16)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.93.5](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.93.4...@times-components/article-main-comment@2.93.5) (2024-12-16)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.93.4](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.93.3...@times-components/article-main-comment@2.93.4) (2024-12-13)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
 ## [2.93.3](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.93.2...@times-components/article-main-comment@2.93.3) (2024-12-12)
 
 **Note:** Version bump only for package @times-components/article-main-comment