+++ packages/related-articles/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.19.5](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.19.4...@times-components/related-articles@6.19.5) (2024-12-16)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
+## [6.19.4](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.19.3...@times-components/related-articles@6.19.4) (2024-12-16)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
+## [6.19.3](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.19.2...@times-components/related-articles@6.19.3) (2024-12-13)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
 ## [6.19.2](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.19.1...@times-components/related-articles@6.19.2) (2024-12-05)
 
 **Note:** Version bump only for package @times-components/related-articles